[Twinkle#67#2] Use int for id instead of long

+++ twinkle-app/twinkle.dal.abstractions/src/main/java/com/github/donkeyrit/twinkle/dal/common/models/Identifiable.java
@@ -1,5 +1,5 @@
 package com.github.donkeyrit.twinkle.dal.common.models;
 
 public interface Identifiable {
-	Long getId();
+	int getId();
 }
